History data table is introduced

+++ db/migrate/20160827112846_create_capitalizable_groups.rb
@@ -1,7 +1,7 @@
 class CreateCapitalizableGroups < ActiveRecord::Migration[5.0]
   def change
     create_table :capitalizable_groups do |t|
-      t.integer :capitalizable_rate
+      t.float :capitalizable_rate
       t.text :description
 
       t.timestamps
